+++ schemas/index.js
@@ -1,0 +1,4 @@
+const contactSchema = require("./contactSchema")
+const updateFavoriteSchema = require('./updateFavoriteSchema')
+
+module.exports = {contactSchema, updateFavoriteSchema}
